refactor(cli): split command orchestration modules
